+++ packages/git-mob-core/src/git-mob-api/git-authors/index.js
@@ -87,17 +87,6 @@ export function gitAuthors(readFilePromise, writeFilePromise, overwriteFilePromi
       return coauthors[authorInitials];
     },
 
-    coAuthorsInitials(authorJson, currentCoAuthors) {
-      const { coauthors } = authorJson;
-      return Object.keys(coauthors).reduce((currentCoAuthorsInitials, initials) => {
-        if (currentCoAuthors.includes(author(coauthors[initials]))) {
-          currentCoAuthorsInitials.push(initials);
-        }
-
-        return currentCoAuthorsInitials;
-      }, []);
-    },
-
     toList(authors) {
       const entries = Object.entries(authors.coauthors);
       return entries.map(([key, { name, email }]) => new Author(key, name, email));