+++ packages/git-mob/src/git-mob-print.ts
@@ -1,9 +1,7 @@
 import os from 'node:os';
 import minimist from 'minimist';
 import { getAllAuthors, getSelectedCoAuthors } from 'git-mob-core';
-import { gitAuthors } from './git-authors/index.js';
 import { runMobPrintHelp } from './helpers.js';
-import { getCoAuthors } from './git-mob-commands.js';
 
 const argv = minimist(process.argv.slice(2), {
   alias: {
@@ -14,15 +12,14 @@ const argv = minimist(process.argv.slice(2), {
 
 await execute(argv);
 
-async function execute(args) {
+async function execute(args: minimist.ParsedArgs) {
   if (args.help) {
     runMobPrintHelp();
     process.exit(0);
   }
 
   if (args.initials) {
-    await printCoAuthorsInitials();
-    process.exit(0);
+    return printCoAuthorsInitials();
   }
 
   return printCoAuthors();
@@ -34,27 +31,24 @@ async function printCoAuthors() {
     const selectedAuthors = getSelectedCoAuthors(allAuthors);
     const coAuthors = selectedAuthors.map(author => author.format()).join(os.EOL);
     console.log(os.EOL + os.EOL + coAuthors);
-  } catch (error) {
-    console.error(`Error: ${error.message}`);
+  } catch (error: unknown) {
+    const printError = error as Error;
+    console.error(`Error: ${printError.message}`);
     process.exit(1);
   }
 }
 
 async function printCoAuthorsInitials() {
   try {
-    const instance = gitAuthors();
-    const authorList = await instance.read();
-    const currentCoAuthors = getCoAuthors();
+    const allAuthors = await getAllAuthors();
+    const selectedAuthors = getSelectedCoAuthors(allAuthors);
 
-    const coAuthorsInitials = instance.coAuthorsInitials(
-      authorList,
-      currentCoAuthors
-    );
-    if (coAuthorsInitials.length > 0) {
-      console.log(coAuthorsInitials.join(','));
+    if (selectedAuthors.length > 0) {
+      console.log(selectedAuthors.map(author => author.key).join(','));
     }
-  } catch (error) {
-    console.error(`Error: ${error.message}`);
+  } catch (error: unknown) {
+    const initialsError = error as Error;
+    console.error(`Error: ${initialsError.message}`);
     process.exit(1);
   }
 }